composite directions
