+++ src/main/java/net/itarray/automotion/internal/geometry/VectorExtendGiving.java
@@ -1,0 +1,46 @@
+package net.itarray.automotion.internal.geometry;
+
+import java.util.function.Function;
+
+import static java.lang.String.format;
+
+public class VectorExtendGiving implements ExtendGiving<Vector> {
+
+    private final Direction x;
+    private final Direction y;
+
+    public VectorExtendGiving(Direction x, Direction y) {
+        this.x = x;
+        this.y = y;
+    }
+
+    @Override
+    public String extendName() {
+        return "size";
+    }
+
+    @Override
+    public String beginName() {
+        return format("%s %s", x.beginName(), y.beginName());
+    }
+
+    @Override
+    public String endName() {
+        return format("%s %s", x.endName(), y.endName());
+    }
+
+    @Override
+    public Vector begin(Rectangle rectangle) {
+        return new Vector(x.begin(rectangle), y.begin(rectangle));
+    }
+
+    @Override
+    public Vector end(Rectangle rectangle) {
+        return new Vector(x.end(rectangle), y.end(rectangle));
+    }
+
+    @Override
+    public Function<Vector, Vector> transform() {
+        return v -> new Vector(x.transform().apply(v), y.transform().apply(v));
+    }
+}
